backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.26.0
Pillow==10.1.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-image==0.22.0
pandas==2.1.4
pytest==7.4.3
